backend/requirements.txt
fastapi
uvicorn
motor
beanie
pydantic-settings
pyjwt
email-validator